Ensure all three primary endpoints have subquery logic
* Update tests for all three endpoints accordingly
* Fix bugs in SQLAlchemy causing issues with relationships
* Refactor some SQLAlchemy models to reduce redundancy
