+++ tests/integration/test_user_profile.py
@@ -4,7 +4,7 @@ import pytest
 
 from middleware.enums import PermissionsEnum
 from middleware.schema_and_dto_logic.common_response_schemas import MessageSchema
-from middleware.schema_and_dto_logic.primary_resource_schemas.data_requests_schemas import (
+from middleware.schema_and_dto_logic.primary_resource_schemas.data_requests_advanced_schemas import (
     GetManyDataRequestsResponseSchema,
 )
 from resources.UserProfile import USER_PROFILE_DATA_REQUEST_ENDPOINT_FULL
