+++ tests/test_schema_aligned_with_db_table.py
@@ -8,9 +8,7 @@ from database_client.database_client import DatabaseClient
 from middleware.enums import Relations
 from middleware.schema_and_dto_logic.primary_resource_schemas.agencies_base_schemas import AgencyInfoBaseSchema, \
     AgenciesExpandedSchema
-from middleware.schema_and_dto_logic.primary_resource_schemas.data_requests_schemas import (
-    DataRequestsSchema,
-)
+from middleware.schema_and_dto_logic.primary_resource_schemas.data_requests_base_schema import DataRequestsSchema
 from tests.conftest import live_database_client
